+++ examples/notification/versions.tf
@@ -2,6 +2,8 @@ terraform {
   required_version = ">= 0.12.6"
 
   required_providers {
-    aws = ">= 3.0"
+    aws    = ">= 3.0"
+    random = ">= 2"
+    null   = ">= 2"
   }
 }