+++ modules/notification/main.tf
@@ -1,8 +1,8 @@
 locals {
   bucket_arn = coalesce(var.bucket_arn, "arn:aws:s3:::${var.bucket}")
 
-  // Convert from "arn:aws:sqs:eu-west-1:835367859851:bold-starling-0" into "https://sqs.eu-west-1.amazonaws.com/835367859851/bold-starling-0" if queue_id was not specified
-  // queue_url used in aws_sqs_queue_policy is not the same as arn which is used in all other places
+  # Convert from "arn:aws:sqs:eu-west-1:835367859851:bold-starling-0" into "https://sqs.eu-west-1.amazonaws.com/835367859851/bold-starling-0" if queue_id was not specified
+  # queue_url used in aws_sqs_queue_policy is not the same as arn which is used in all other places
   queue_ids = { for k, v in var.sqs_notifications : k => format("https://%s.%s.amazonaws.com/%s/%s", data.aws_arn.queue[k].service, data.aws_arn.queue[k].region, data.aws_arn.queue[k].account, data.aws_arn.queue[k].resource) if lookup(v, "queue_id", "") == "" }
 }
 
@@ -54,7 +54,7 @@ resource "aws_s3_bucket_notification" "this" {
   ]
 }
 
-// Lambda
+# Lambda
 resource "aws_lambda_permission" "allow" {
   for_each = var.lambda_notifications
 
@@ -66,7 +66,7 @@ resource "aws_lambda_permission" "allow" {
   source_arn          = local.bucket_arn
 }
 
-// SQS Queue
+# SQS Queue
 data "aws_arn" "queue" {
   for_each = var.sqs_notifications
 
@@ -74,7 +74,7 @@ data "aws_arn" "queue" {
 }
 
 data "aws_iam_policy_document" "sqs" {
-  for_each = var.sqs_notifications
+  for_each = var.create_sqs_policy ? var.sqs_notifications : tomap({})
 
   statement {
     sid = "AllowSQSS3BucketNotification"
@@ -101,15 +101,15 @@ data "aws_iam_policy_document" "sqs" {
 }
 
 resource "aws_sqs_queue_policy" "allow" {
-  for_each = var.sqs_notifications
+  for_each = var.create_sqs_policy ? var.sqs_notifications : tomap({})
 
   queue_url = lookup(each.value, "queue_id", lookup(local.queue_ids, each.key, null))
   policy    = data.aws_iam_policy_document.sqs[each.key].json
 }
 
-// SNS Topic
+# SNS Topic
 data "aws_iam_policy_document" "sns" {
-  for_each = var.sns_notifications
+  for_each = var.create_sns_policy ? var.sns_notifications : tomap({})
 
   statement {
     sid = "AllowSNSS3BucketNotification"
@@ -136,7 +136,7 @@ data "aws_iam_policy_document" "sns" {
 }
 
 resource "aws_sns_topic_policy" "allow" {
-  for_each = var.sns_notifications
+  for_each = var.create_sns_policy ? var.sns_notifications : tomap({})
 
   arn    = each.value.topic_arn
   policy = data.aws_iam_policy_document.sns[each.key].json